requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
fastapi
uvicorn
sse_starlette
bs4
pypdf
groq
chromadb
faiss-cpu
langchain_groq
cassio
beautifulsoup4
langchain-groq
unstructured
pytube
validators==0.28.1
youtube_transcript_api